Changed all instances of 'TestPage1' (or Testpage) to 'HomePage' for better readability
