+++ .vscode/settings.json
@@ -4,5 +4,8 @@
   "prettier.requireConfig": true,
   "[prisma]": {
     "editor.defaultFormatter": "Prisma.prisma"
+  },
+  "[typescriptreact]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
   }
 }
