+++ src/components/Modules/HomePage/HomePageHeaderModule.tsx
@@ -1,0 +1,9 @@
+import { HeaderAtom } from "~/components/Atoms/HeaderAtom/HeaderAtom";
+
+type Props = {
+  headerProps: React.ComponentProps<typeof HeaderAtom>;
+};
+
+export const HomePageHeaderModule = (props: Props) => {
+  return <HeaderAtom {...props.headerProps} />;
+};
