--- src/components/Modules/TestPage1/TestPage1FooterModule.tsx
@@ -1,13 +1,0 @@
-import { IconFooterAtom } from "~/components/Atoms/FooterAtom/IconFooterAtom";
-
-type Props = {
-    footerProps: React.ComponentProps<typeof IconFooterAtom>;
-};
-
-export const TestPage1FooterModule = (props: Props) => {
-    return (
-        <>
-            <IconFooterAtom {...props.footerProps} />
-        </>
-    );
-};
