--- src/components/Modules/TestPage1/TestPage1HeaderModule.tsx
@@ -1,9 +1,0 @@
-import { HeaderAtom } from "~/components/Atoms/HeaderAtom/HeaderAtom"
-
-type Props = {
-    headerProps: React.ComponentProps<typeof HeaderAtom>;
-};
-
-export const TestPage1HeaderModule = (props: Props) => {
-    return <HeaderAtom {...props.headerProps} />;
-};
